Added IsInGroup as IsAdminInGroup checks to MVC controllers. todo same in api controllers

+++ CarsWebServer/WebApp/Controllers/CarInsurancesController.cs
@@ -69,6 +69,12 @@ public class CarInsurancesController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> Create(CarInsurance carInsurance)
     {
+        var isAdmin = await _uow.GroupRepository.IsUserInGroup(User.GetUserId(), carInsurance.Car!.GroupId);
+        if (!isAdmin)
+        {
+            return Forbid();
+        }
+        
         if (ModelState.IsValid)
         {
             carInsurance.EndDate = DateTime.SpecifyKind(carInsurance.EndDate, DateTimeKind.Utc);
@@ -99,6 +105,13 @@ public class CarInsurancesController : Controller
         {
             return NotFound();
         }
+        
+        var isAdmin = await _uow.GroupRepository.IsUserAdminInGroup(User.GetUserId(), entity.Car!.GroupId);
+        if (!isAdmin)
+        {
+            return Forbid();
+        }
+        
         var userGroups = await _uow.GroupRepository.AllAsync(User.GetUserId());
         var userCars = await _uow.CarRepository.AllCarsAsync(userGroups);
 
@@ -117,6 +130,12 @@ public class CarInsurancesController : Controller
         if (id != carInsurance.Id)
         {
             return NotFound();
+        }
+        
+        var isAdmin = await _uow.GroupRepository.IsUserAdminInGroup(User.GetUserId(), carInsurance.Car!.GroupId);
+        if (!isAdmin)
+        {
+            return Forbid();
         }
 
         if (ModelState.IsValid)
@@ -145,6 +164,12 @@ public class CarInsurancesController : Controller
         {
             return NotFound();
         }
+        
+        var isAdmin = await _uow.GroupRepository.IsUserAdminInGroup(User.GetUserId(), entity.Car!.GroupId);
+        if (!isAdmin)
+        {
+            return Forbid();
+        }
 
         return View(entity);
     }
@@ -154,6 +179,19 @@ public class CarInsurancesController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(Guid id)
     {
+        var entity = await _uow.CarInsuranceRepository.FindAsync(id, User.GetUserId());
+        if (entity == null)
+        {
+            return NotFound();
+        }
+
+        // Check if current user is admin of the group
+        var isAdmin = await _uow.GroupRepository.IsUserAdminInGroup(User.GetUserId(), entity.Car!.GroupId);
+        if (!isAdmin)
+        {
+            return Forbid();
+        }
+        
         await _uow.CarInsuranceRepository.RemoveAsync(id);
         await _uow.SaveChangesAsync();
         return RedirectToAction(nameof(Index));
